move PinBox into pin module and export through std
