+++ src/liballoc/lib.rs
@@ -159,6 +159,7 @@ pub mod collections;
 pub mod sync;
 pub mod rc;
 pub mod raw_vec;
+pub mod pin;
 pub mod prelude;
 pub mod borrow;
 pub mod fmt;
