+++ src/test/run-pass/async-await.rs
@@ -12,7 +12,7 @@
 
 #![feature(arbitrary_self_types, async_await, await_macro, futures_api, pin)]
 
-use std::boxed::PinBox;
+use std::pin::PinBox;
 use std::pin::PinMut;
 use std::future::Future;
 use std::sync::{
